Implemented new esse3 api and refactored data-api modules

+++ android/data-api/bicoccapp/src/test/kotlin/it/attendance100/mybicocca/data/api/bicoccapp/BicoccappAuthSession.kt
@@ -5,5 +5,5 @@ data class BicoccappAuthSession(
     val client: String,
     val uid: String,
     val fiscalCode: String,
-    val matricId: Int
+    val matricId: Long
 )
